Extract client method for updates in UpdateClient class
