+++ tests/clients/BarracksClient.spec.js
@@ -49,96 +49,21 @@ describe('Barracks', () => {
       expect(barracksClient).to.have.property('getAccount').and.to.be.a('function');
       expect(barracksClient).to.have.property('setGoogleAnalyticsTrackingId').and.to.be.a('function');
     });
-  });
 
-  describe('#editUpdate()', () => {
+    it('should initialize methods from updateClient when constructor called', () => {
+      // When
+      const barracksClient = new Barracks();
 
-    const originalUpdate = {
-      name: 'test update',
-      versionId: '2.3',
-      uuid: '123456789',
-      packageInfo: { id: '123456' }
-    };
-
-    const originalUpdateWithPackageId = {
-      name: 'test update',
-      versionId: '2.3',
-      uuid: '123456789',
-      packageId: '123456'
-    };
-
-    it('should return an error message when edit request fails', done => {
-      // Given
-      const changes = { uuid: originalUpdate.uuid, name: 'test update' };
-      const error = { message: 'Error !' };
-      barracks.getUpdate = sinon.stub().returns(Promise.resolve(originalUpdate));
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(error));
-
-      // When / Then
-      barracks.editUpdate(token, changes).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(error.message);
-        expect(barracks.getUpdate).to.have.been.calledOnce;
-        expect(barracks.getUpdate).to.have.been.calledWithExactly(token, changes.uuid);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('editUpdate', {
-          headers: { 'x-auth-token': token },
-          body: Object.assign({}, originalUpdateWithPackageId, changes),
-          pathVariables: { uuid: originalUpdateWithPackageId.uuid }
-        });
-        done();
-      });
+      // Then
+      expect(barracksClient).to.have.property('createUpdate').and.to.be.a('function');
+      expect(barracksClient).to.have.property('editUpdate').and.to.be.a('function');
+      expect(barracksClient).to.have.property('getUpdate').and.to.be.a('function');
+      expect(barracksClient).to.have.property('getUpdates').and.to.be.a('function');
+      expect(barracksClient).to.have.property('getUpdatesBySegmentId').and.to.be.a('function');
+      expect(barracksClient).to.have.property('publishUpdate').and.to.be.a('function');
+      expect(barracksClient).to.have.property('archiveUpdate').and.to.be.a('function');
+      expect(barracksClient).to.have.property('scheduleUpdate').and.to.be.a('function');
     });
-
-    it('should return an error message when the update does not exist', done => {
-      // Given
-      const changes = { uuid: originalUpdate.uuid, name: 'test update' };
-      const error = { message: 'Error !' };
-      const excpectedResult = Object.assign({}, originalUpdateWithPackageId, changes);
-      barracks.getUpdate = sinon.stub().returns(Promise.reject(error));
-      barracks.client.sendEndpointRequest = sinon.spy();
-
-      // When / Then
-      barracks.editUpdate(token, changes).then(result => {
-        done('Should have failed');
-      }).catch(err => {
-        try {
-          expect(err).to.be.equals(error.message);
-          expect(barracks.getUpdate).to.have.been.calledOnce;
-          expect(barracks.getUpdate).to.have.been.calledWithExactly(token, changes.uuid);
-          expect(barracks.client.sendEndpointRequest).to.not.have.been.called;
-          done();
-        } catch (e) {
-          done(e);
-        }
-      });
-    });
-
-    it('should return the edited update', done => {
-      // Given
-      const changes = { uuid: originalUpdate.uuid, name: 'test update' };
-      const response = { body: originalUpdate };
-      barracks.getUpdate = sinon.stub().returns(Promise.resolve(originalUpdate));
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-      const excpectedResult = Object.assign({}, originalUpdateWithPackageId, changes);
-
-      // When / Then
-      barracks.editUpdate(token, changes).then(result => {
-        expect(barracks.getUpdate).to.have.been.calledOnce;
-        expect(barracks.getUpdate).to.have.been.calledWithExactly(token, changes.uuid);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('editUpdate', {
-          headers: { 'x-auth-token': token },
-          body: excpectedResult,
-          pathVariables: { uuid: originalUpdateWithPackageId.uuid }
-        });
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-
   });
 
   describe('#editSegment()', () => {
@@ -275,256 +200,6 @@ describe('Barracks', () => {
         done();
       }).catch(err => {
         done(err);
-      });
-    });
-
-  });
-
-  describe('#getUpdate()', () => {
-
-    const existingUpdate = {
-      uuid: '123456789',
-      name: 'Plop',
-      versionId: 'Cool'
-    };
-
-    it('should return an error message when request fails', done => {
-      // Given
-      const error = { message: 'Error !' };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(error));
-
-      // When / Then
-      barracks.getUpdate(token, existingUpdate.uuid).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(error.message);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('getUpdate', {
-          headers: { 'x-auth-token': token },
-          pathVariables: {
-            uuid: existingUpdate.uuid
-          }
-        });
-        done();
-      });
-    });
-
-    it('should return an update when request succeed', done => {
-      // Given
-      const response = { body: existingUpdate };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.getUpdate(token, existingUpdate.uuid).then(result => {
-        expect(result).to.be.equals(existingUpdate);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('getUpdate', {
-          headers: { 'x-auth-token': token },
-          pathVariables: {
-            uuid: existingUpdate.uuid
-          }
-        });
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-
-  });
-
-  describe('#getUpdates()', () => {
-
-    it('should return a stream object and deleguate to the client', done => {
-      // Given
-      const options = {
-        headers: {
-          'x-auth-token': token
-        }
-      }
-      barracks.client.retrieveAllPages = sinon.spy();
-
-      // When / Then
-      barracks.getUpdates(token).then(result => {
-        expect(result).to.be.instanceOf(PageableStream);
-        expect(barracks.client.retrieveAllPages).to.have.been.calledOnce;
-        expect(barracks.client.retrieveAllPages).to.have.been.calledWithExactly(
-          new PageableStream(),
-          'getUpdates',
-          options,
-          'updates'
-        );
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-  });
-
-  describe('#getUpdatesySegmentId()', () => {
-
-    it('should return a stream object and deleguate to the client', done => {
-      // Given
-      const segmentId = 'mySegment';
-      const options = {
-        headers: { 'x-auth-token': token },
-        pathVariables: { segmentId }
-      };
-      barracks.client.retrieveAllPages = sinon.spy();
-
-      // When / Then
-      barracks.getUpdatesBySegmentId(token, segmentId).then(result => {
-        expect(result).to.be.instanceOf(PageableStream);
-        expect(barracks.client.retrieveAllPages).to.have.been.calledOnce;
-        expect(barracks.client.retrieveAllPages).to.have.been.calledWithExactly(
-          new PageableStream(),
-          'updatesBySegmentId',
-          options,
-          'updates'
-        );
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-  });
-
-  describe('#publishUpdate()', () => {
-
-    it('should return an error message when request fails', done => {
-      // Given
-      const updateId = 'poiuytrewq';
-      const error = { message: 'Error !' };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(error));
-
-      // When / Then
-      barracks.publishUpdate(token, updateId).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(error.message);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('publishUpdate', {
-          headers: { 'x-auth-token': token },
-          pathVariables: { uuid: updateId }
-        });
-        done();
-      });
-    });
-
-    it('should return update info when publication succeed', done => {
-      // Given
-      const updateId = 'poiuytrewq';
-      const update = { uuid: updateId, status: 'published' };
-      const response = { body: update };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.publishUpdate(token, updateId).then(result => {
-        expect(result).to.be.equals(update);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('publishUpdate', {
-          headers: { 'x-auth-token': token },
-          pathVariables: { uuid: updateId }
-        });
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-  });
-
-  describe('#archiveUpdate()', () => {
-
-    it('should return an error message when request fails', done => {
-      // Given
-      const updateId = 'poiuytrewq';
-      const error = { message: 'Error !' };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(error));
-
-      // When / Then
-      barracks.archiveUpdate(token, updateId).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(error.message);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('archiveUpdate', {
-          headers: { 'x-auth-token': token },
-          pathVariables: { uuid: updateId }
-        });
-        done();
-      });
-    });
-
-    it('should return update info when archivage succeed', done => {
-      // Given
-      const updateId = 'poiuytrewq';
-      const update = { uuid: updateId, status: 'archived' };
-      const response = { body: update };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.archiveUpdate(token, updateId).then(result => {
-        expect(result).to.be.equals(update);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('archiveUpdate', {
-          headers: { 'x-auth-token': token },
-          pathVariables: { uuid: updateId }
-        });
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-  });
-
-  describe('#scheduleUpdate()', () => {
-
-    it('should return update info when schedule succeed', done => {
-      // Given
-      const uuid = 'poiuytrewq';
-      const date = new Date();
-      const update = { uuid: uuid, status: 'scheduled' };
-      const response = { body: update };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.scheduleUpdate(token, uuid, date).then(result => {
-        expect(result).to.be.equals(update);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('scheduleUpdate', {
-          headers: { 'x-auth-token': token },
-          pathVariables: {
-            uuid,
-            time: date.toISOString()
-          }
-        });
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-
-    it('should return an error when schedule failed', done => {
-      // Given
-      const uuid = 'poiuytrewq';
-      const date = new Date();
-      const error = 'Error!';
-      const response = { message: error };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(response));
-
-      // When / Then
-      barracks.scheduleUpdate(token, uuid, date).then(result => {
-        done('Should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(error);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('scheduleUpdate', {
-          headers: { 'x-auth-token': token },
-          pathVariables: {
-            uuid,
-            time: date.toISOString()
-          }
-        });
-        done();
       });
     });
   });
@@ -843,51 +518,6 @@ describe('Barracks', () => {
         done(err);
       });
     });
-  });
-
-  describe('#createUpdate()', () => {
-
-    it('should return an error message when request fails', done => {
-      // Given
-      const update = { name: 'test update', versionId: '2.3' };
-      const error = { message: 'Error !' };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(error));
-
-      // When / Then
-      barracks.createUpdate(token, update).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(error.message);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('createUpdate', {
-          headers: { 'x-auth-token': token },
-          body: update
-        });
-        done();
-      });
-    });
-
-    it('should return the update created', done => {
-      // Given
-      const update = { name: 'test update', versionId: '2.3' };
-      const updateFull = { name: 'test update', versionId: '2.3', uuid: '1234567890poiuytrew' };
-      const response = { body: updateFull };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.createUpdate(token, update).then(result => {
-        expect(result).to.be.equals(updateFull);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('createUpdate', {
-          headers: { 'x-auth-token': token },
-          body: update
-        });
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-
   });
 
   describe('#createSegment()', () => {
